Bet Ajax Request Added

+++ db/migrate/20231010202242_add_outcome_to_bets.rb
@@ -1,5 +1,5 @@
 class AddOutcomeToBets < ActiveRecord::Migration[7.1]
   def change
-    add_column :bets, :outcome, :float
+    add_column :bets, :outcome, :decimal
   end
 end
